Extract JavaMarshal registered peer tracking
Move the JavaMarshal peer-registration and GC bridge bookkeeping out of JavaMarshalValueManager into a reusable JavaMarshalRegisteredPeers helper, and share the trimmable peer-type/cast helpers.
