+++ src/Mono.Android/Microsoft.Android.Runtime/JavaMarshalValueManager.cs
@@ -1,18 +1,7 @@
-// Originally from: https://github.com/dotnet/java-interop/blob/9b1d8781e8e322849d05efac32119c913b21c192/src/Java.Runtime.Environment/Java.Interop/ManagedValueManager.cs
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
-using System.IO;
-using System.Linq;
 using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Runtime.InteropServices.Java;
-using System.Threading;
 using Android.Runtime;
 using Java.Interop;
 
@@ -21,30 +10,16 @@ namespace Microsoft.Android.Runtime;
 class JavaMarshalValueManager : AndroidReflectionJniValueManager
 {
 	const DynamicallyAccessedMemberTypes Constructors = DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors;
+	const BindingFlags ActivationConstructorBindingFlags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance;
 
-	readonly Dictionary<int, List<ReferenceTrackingHandle>> RegisteredInstances = new ();
-	readonly ConcurrentQueue<IntPtr> CollectedContexts = new ();
+	static readonly Type[] XAConstructorSignature = new Type [] { typeof (IntPtr), typeof (JniHandleOwnership) };
 
-	bool disposed;
-
-	public unsafe JavaMarshalValueManager ()
-	{
-		var javaMarshalValueManagerHandle = new GCHandle<JavaMarshalValueManager> (this);
-		var mark_cross_references_ftn = RuntimeNativeMethods.clr_initialize_gc_bridge (
-			GCHandle<JavaMarshalValueManager>.ToIntPtr (javaMarshalValueManagerHandle), &BridgeProcessingStarted, &BridgeProcessingFinished);
-		JavaMarshal.Initialize (mark_cross_references_ftn);
-	}
+	readonly JavaMarshalRegisteredPeers registeredPeers = new ();
 
 	protected override void Dispose (bool disposing)
 	{
-		disposed = true;
+		registeredPeers.Dispose ();
 		base.Dispose (disposing);
-	}
-
-	void ThrowIfDisposed ()
-	{
-		if (disposed)
-			throw new ObjectDisposedException (nameof (JavaMarshalValueManager));
 	}
 
 	public override void WaitForGCBridgeProcessing ()
@@ -57,189 +32,29 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 		// they are not affected by the bridge swapping control_block handles.
 	}
 
-	public unsafe override void CollectPeers ()
+	public override void CollectPeers ()
 	{
-		ThrowIfDisposed ();
-
-		while (CollectedContexts.TryDequeue (out IntPtr contextPtr)) {
-			Debug.Assert (contextPtr != IntPtr.Zero, "CollectedContexts should not contain null pointers.");
-			HandleContext* context = (HandleContext*)contextPtr;
-			
-			lock (RegisteredInstances) {
-				Remove (context);
-			}
-
-			HandleContext.Free (ref context);
-		}
-
-		void Remove (HandleContext* context)
-		{
-			int key = context->PeerIdentityHashCode;
-			if (!RegisteredInstances.TryGetValue (key, out List<ReferenceTrackingHandle>? peers))
-				return;
-
-			for (int i = peers.Count - 1; i >= 0; i--) {
-				var peer = peers [i];
-				if (peer.BelongsToContext (context)) {
-					peers.RemoveAt (i);
-				}
-			}
-
-			if (peers.Count == 0) {
-				RegisteredInstances.Remove (key);
-			}
-		}
+		registeredPeers.CollectPeers ();
 	}
 
 	public override void AddPeer (IJavaPeerable value)
 	{
-		ThrowIfDisposed ();
-
-		// Remove any collected contexts before adding a new peer.
-		CollectPeers ();
-
-		var r = value.PeerReference;
-		if (!r.IsValid)
-			throw new ObjectDisposedException (value.GetType ().FullName);
-
-		if (r.Type != JniObjectReferenceType.Global) {
-			value.SetPeerReference (r.NewGlobalRef ());
-			JniObjectReference.Dispose (ref r, JniObjectReferenceOptions.CopyAndDispose);
-		}
-		int key = value.JniIdentityHashCode;
-		lock (RegisteredInstances) {
-			List<ReferenceTrackingHandle>? peers;
-			if (!RegisteredInstances.TryGetValue (key, out peers)) {
-				peers = [new ReferenceTrackingHandle (value)];
-				RegisteredInstances.Add (key, peers);
-				return;
-			}
-
-			for (int i = peers.Count - 1; i >= 0; i--) {
-				ReferenceTrackingHandle peer = peers [i];
-				if (peer.Target is not IJavaPeerable target)
-					continue;
-				if (!JniEnvironment.Types.IsSameObject (target.PeerReference, value.PeerReference))
-					continue;
-				if (target.JniManagedPeerState.HasFlag (JniManagedPeerStates.Replaceable)) {
-					peer.Dispose ();
-					peers [i] = new ReferenceTrackingHandle (value);
-				} else {
-					WarnNotReplacing (key, value, target);
-				}
-				GC.KeepAlive (target);
-				return;
-			}
-
-			peers.Add (new ReferenceTrackingHandle (value));
-		}
-	}
-
-	void WarnNotReplacing (int key, IJavaPeerable ignoreValue, IJavaPeerable keepValue)
-	{
-		Runtime.ObjectReferenceManager.WriteGlobalReferenceLine (
-				"Warning: Not registering PeerReference={0} IdentityHashCode=0x{1} Instance={2} Instance.Type={3} Java.Type={4}; " +
-				"keeping previously registered PeerReference={5} Instance={6} Instance.Type={7} Java.Type={8}.",
-				ignoreValue.PeerReference.ToString (),
-				key.ToString ("x", CultureInfo.InvariantCulture),
-				RuntimeHelpers.GetHashCode (ignoreValue).ToString ("x", CultureInfo.InvariantCulture),
-				ignoreValue.GetType ().FullName,
-				JniEnvironment.Types.GetJniTypeNameFromInstance (ignoreValue.PeerReference),
-				keepValue.PeerReference.ToString (),
-				RuntimeHelpers.GetHashCode (keepValue).ToString ("x", CultureInfo.InvariantCulture),
-				keepValue.GetType ().FullName,
-				JniEnvironment.Types.GetJniTypeNameFromInstance (keepValue.PeerReference));
+		registeredPeers.AddPeer (value);
 	}
 
 	public override IJavaPeerable? PeekPeer (JniObjectReference reference)
 	{
-		ThrowIfDisposed ();
-
-		if (!reference.IsValid)
-			return null;
-
-		int key = GetJniIdentityHashCode (reference);
-
-		lock (RegisteredInstances) {
-			if (!RegisteredInstances.TryGetValue (key, out List<ReferenceTrackingHandle>? peers))
-				return null;
-
-			for (int i = peers.Count - 1; i >= 0; i--) {
-				if (peers [i].Target is IJavaPeerable peer
-					&& JniEnvironment.Types.IsSameObject (reference, peer.PeerReference))
-				{
-					return peer;
-				}
-			}
-
-			if (peers.Count == 0)
-				RegisteredInstances.Remove (key);
-		}
-		return null;
+		return registeredPeers.PeekPeer (reference);
 	}
 
 	public override void RemovePeer (IJavaPeerable value)
 	{
-		ThrowIfDisposed ();
-
-		// Remove any collected contexts before modifying RegisteredInstances
-		CollectPeers ();
-
-		if (value == null)
-			throw new ArgumentNullException (nameof (value));
-
-		lock (RegisteredInstances) {
-			int key = value.JniIdentityHashCode;
-			if (!RegisteredInstances.TryGetValue (key, out List<ReferenceTrackingHandle>? peers))
-				return;
-
-			for (int i = peers.Count - 1; i >= 0; i--) {
-				ReferenceTrackingHandle peer = peers [i];
-				IJavaPeerable target = peer.Target;
-				if (ReferenceEquals (value, target)) {
-					peers.RemoveAt (i);
-					peer.Dispose ();
-				}
-				GC.KeepAlive (target);
-			}
-			if (peers.Count == 0)
-				RegisteredInstances.Remove (key);
-		}
+		registeredPeers.RemovePeer (value);
 	}
 
 	public override void FinalizePeer (IJavaPeerable value)
 	{
-		var h = value.PeerReference;
-		var o = Runtime.ObjectReferenceManager;
-		// MUST NOT use SafeHandle.ReferenceType: local refs are tied to a JniEnvironment
-		// and the JniEnvironment's corresponding thread; it's a thread-local value.
-		// Accessing SafeHandle.ReferenceType won't kill anything (so far...), but
-		// instead it always returns JniReferenceType.Invalid.
-		if (!h.IsValid || h.Type == JniObjectReferenceType.Local) {
-			if (o.LogGlobalReferenceMessages) {
-				o.WriteGlobalReferenceLine ("Finalizing PeerReference={0} IdentityHashCode=0x{1} Instance=0x{2} Instance.Type={3}",
-						h.ToString (),
-						value.JniIdentityHashCode.ToString ("x", CultureInfo.InvariantCulture),
-						RuntimeHelpers.GetHashCode (value).ToString ("x", CultureInfo.InvariantCulture),
-						value.GetType ().ToString ());
-			}
-			RemovePeer (value);
-			value.SetPeerReference (new JniObjectReference ());
-			value.Finalized ();
-			return;
-		}
-
-		RemovePeer (value);
-		if (o.LogGlobalReferenceMessages) {
-			o.WriteGlobalReferenceLine ("Finalizing PeerReference={0} IdentityHashCode=0x{1} Instance=0x{2} Instance.Type={3}",
-					h.ToString (),
-					value.JniIdentityHashCode.ToString ("x", CultureInfo.InvariantCulture),
-					RuntimeHelpers.GetHashCode (value).ToString ("x", CultureInfo.InvariantCulture),
-					value.GetType ().ToString ());
-		}
-		value.SetPeerReference (new JniObjectReference ());
-		JniObjectReference.Dispose (ref h);
-		value.Finalized ();
+		registeredPeers.FinalizePeer (value);
 	}
 
 	public override void ActivatePeer (JniObjectReference reference, [DynamicallyAccessedMembers (Constructors)] Type type, ConstructorInfo cinfo, object?[]? argumentValues)
@@ -252,241 +67,8 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 
 	public override List<JniSurfacedPeerInfo> GetSurfacedPeers ()
 	{
-		ThrowIfDisposed ();
-
-		// Remove any collected contexts before iterating over all the registered instances
-		CollectPeers ();
-
-		lock (RegisteredInstances) {
-			var peers = new List<JniSurfacedPeerInfo> (RegisteredInstances.Count);
-			foreach (var (identityHashCode, referenceTrackingHandles) in RegisteredInstances) {
-				foreach (var peer in referenceTrackingHandles) {
-					if (peer.Target is IJavaPeerable target) {
-						peers.Add (new JniSurfacedPeerInfo (identityHashCode, new WeakReference<IJavaPeerable> (target)));
-					}
-				}
-			}
-			return peers;
-		}
+		return registeredPeers.GetSurfacedPeers ();
 	}
-
-	unsafe struct ReferenceTrackingHandle : IDisposable
-	{
-		WeakReference<IJavaPeerable> _weakReference;
-		HandleContext* _context;
-
-		public bool BelongsToContext (HandleContext* context)
-			=> _context == context;
-
-		public ReferenceTrackingHandle (IJavaPeerable peer)
-		{
-			_context = HandleContext.Alloc (peer);
-			_weakReference = new WeakReference<IJavaPeerable> (peer);
-		}
-
-		public IJavaPeerable? Target
-			=> _weakReference.TryGetTarget (out var target) ? target : null;
-
-		public void Dispose ()
-		{
-			if (_context == null)
-				return;
-
-			IJavaPeerable? target = Target;
-
-			GCHandle handle = HandleContext.GetAssociatedGCHandle (_context);
-			HandleContext.Free (ref _context);
-			_weakReference.SetTarget (null);
-			if (handle.IsAllocated) {
-				handle.Free ();
-			}
-
-			// Make sure the target is not collected before we finish disposing
-			GC.KeepAlive (target);
-		}
-	}
-
-	[StructLayout (LayoutKind.Sequential)]
-	unsafe struct HandleContext
-	{
-		static readonly nuint Size = (nuint)Marshal.SizeOf<HandleContext> ();
-		static readonly Dictionary<IntPtr, GCHandle> referenceTrackingHandles = new ();
-
-		int identityHashCode;
-		IntPtr controlBlock;
-
-		public int PeerIdentityHashCode => identityHashCode;
-		public bool IsCollected
-		{
-			get
-			{
-				if (controlBlock == IntPtr.Zero)
-					throw new InvalidOperationException ("HandleContext control block is not initialized.");
-
-				return ((JniObjectReferenceControlBlock*) controlBlock)->handle == IntPtr.Zero;
-			}
-		}
-
-		// This is an internal mirror of the Java.Interop.JniObjectReferenceControlBlock
-		private struct JniObjectReferenceControlBlock
-		{
-			public IntPtr handle;
-			public int handle_type;
-			public IntPtr weak_handle;
-			public int refs_added;
-		}
-
-		public static GCHandle GetAssociatedGCHandle (HandleContext* context)
-		{
-			lock (referenceTrackingHandles) {
-				if (!referenceTrackingHandles.TryGetValue ((IntPtr) context, out GCHandle handle)) {
-					throw new InvalidOperationException ("Unknown reference tracking handle.");
-				}
-
-				return handle;
-			}
-		}
-
-		public static unsafe void EnsureAllContextsAreOurs (MarkCrossReferencesArgs* mcr)
-		{
-// This call site is reachable on all platforms. 'MarkCrossReferencesArgs.ComponentCount' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'MarkCrossReferencesArgs.Components' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'StronglyConnectedComponent.Count' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'StronglyConnectedComponent.Contexts' is only supported on: 'android'.
-#pragma warning disable CA1416
-
-			lock (referenceTrackingHandles) {
-				for (nuint i = 0; i < mcr->ComponentCount; i++) {
-					StronglyConnectedComponent component = mcr->Components [i];
-					EnsureAllContextsInComponentAreOurs (component);
-				}
-			}
-
-			static void EnsureAllContextsInComponentAreOurs (StronglyConnectedComponent component)
-			{
-				for (nuint i = 0; i < component.Count; i++) {
-					EnsureContextIsOurs ((IntPtr)component.Contexts [i]);
-				}
-			}
-
-			static void EnsureContextIsOurs (IntPtr context)
-			{
-				if (!referenceTrackingHandles.ContainsKey (context)) {
-					throw new InvalidOperationException ("Unknown reference tracking handle.");
-				}
-			}
-
-#pragma warning restore CA1416
-		}
-
-		public static HandleContext* Alloc (IJavaPeerable peer)
-		{
-			var context = (HandleContext*) NativeMemory.AllocZeroed (1, Size);
-			if (context == null) {
-				throw new OutOfMemoryException ("Failed to allocate memory for HandleContext.");
-			}
-
-			context->identityHashCode = peer.JniIdentityHashCode;
-			context->controlBlock = peer.JniObjectReferenceControlBlock;
-
-			GCHandle handle = JavaMarshal.CreateReferenceTrackingHandle (peer, context);
-			lock (referenceTrackingHandles) {
-				referenceTrackingHandles [(IntPtr) context] = handle;
-			}
-
-			return context;
-		}
-
-		public static void Free (ref HandleContext* context)
-		{
-			if (context == null) {
-				return;
-			}
-
-			lock (referenceTrackingHandles) {
-				referenceTrackingHandles.Remove ((IntPtr)context);
-			}
-
-			NativeMemory.Free (context);
-			context = null;
-		}
-	}
-
-	[UnmanagedCallersOnly]
-	static unsafe void BridgeProcessingStarted (MarkCrossReferencesArgs* mcr)
-	{
-		if (mcr == null) {
-			throw new ArgumentNullException (nameof (mcr), "MarkCrossReferencesArgs should never be null.");
-		}
-
-		HandleContext.EnsureAllContextsAreOurs (mcr);
-	}
-
-	[UnmanagedCallersOnly]
-	static unsafe void BridgeProcessingFinished (IntPtr javaMarshalValueManagerHandle, MarkCrossReferencesArgs* mcr)
-	{
-		if (mcr == null) {
-			throw new ArgumentNullException (nameof (mcr), "MarkCrossReferencesArgs should never be null.");
-		}
-
-		JavaMarshalValueManager instance = GCHandle<JavaMarshalValueManager>.FromIntPtr (javaMarshalValueManagerHandle).Target;
-
-		ReadOnlySpan<GCHandle> handlesToFree = instance.ProcessCollectedContexts (mcr);
-
-
-// This call site is reachable on all platforms. 'JavaMarshal.FinishCrossReferenceProcessing(MarkCrossReferencesArgs*, ReadOnlySpan<GCHandle>)' is only supported on: 'android'.
-#pragma warning disable CA1416
-		JavaMarshal.FinishCrossReferenceProcessing (mcr, handlesToFree);
-#pragma warning restore CA1416
-	}
-
-	unsafe ReadOnlySpan<GCHandle> ProcessCollectedContexts (MarkCrossReferencesArgs* mcr)
-	{
-		List<GCHandle> handlesToFree = [];
-
-// This call site is reachable on all platforms. 'MarkCrossReferencesArgs.ComponentCount' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'MarkCrossReferencesArgs.Components' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'StronglyConnectedComponent.Count' is only supported on: 'android'.
-// This call site is reachable on all platforms. 'StronglyConnectedComponent.Contexts' is only supported on: 'android'.
-#pragma warning disable CA1416
-
-		for (int i = 0; (nuint)i < mcr->ComponentCount; i++) {
-			StronglyConnectedComponent component = mcr->Components [i];
-			for (int j = 0; (nuint)j < component.Count; j++) {
-				ProcessContext ((HandleContext*)component.Contexts [j]);
-			}
-		}
-
-#pragma warning restore CA1416
-
-		void ProcessContext (HandleContext* context)
-		{
-			if (context == null) {
-				throw new ArgumentNullException (nameof (context), "HandleContext should never be null.");
-			}
-
-			// Ignore contexts which were not collected
-			if (!context->IsCollected) {
-				return;
-			}
-
-			GCHandle handle = HandleContext.GetAssociatedGCHandle (context);
-
-			// Note: modifying the RegisteredInstances dictionary while processing the collected contexts
-			// is tricky and can lead to deadlocks, so we remember which contexts were collected and we will free
-			// them later outside of the bridge processing loop.
-			CollectedContexts.Enqueue ((IntPtr)context);
-
-			// important: we must not free the handle before passing it to JavaMarshal.FinishCrossReferenceProcessing
-			handlesToFree.Add (handle);
-		}
-
-		return CollectionsMarshal.AsSpan (handlesToFree);
-	}
-
-	const BindingFlags ActivationConstructorBindingFlags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance;
-
-	static  readonly    Type[]  XAConstructorSignature  = new Type [] { typeof (IntPtr), typeof (JniHandleOwnership) };
 
 	public override IJavaPeerable? CreatePeer (
 			ref JniObjectReference reference,
@@ -494,7 +76,7 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 			[DynamicallyAccessedMembers (Constructors)]
 			Type? targetType)
 	{
-		ThrowIfDisposed ();
+		EnsureNotDisposed ();
 
 		if (!reference.IsValid) {
 			return null;
@@ -505,7 +87,7 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 				// Mirror legacy GetPeerType: callers commonly request universal
 				// interfaces / boxes (IJavaPeerable, object, Exception) — map these
 				// to a concrete peer type so the proxy lookup can succeed.
-				var resolvedTargetType = ResolvePeerType (targetType);
+				var resolvedTargetType = JavaMarshalValueManagerHelper.ResolvePeerType (targetType);
 
 				var typeMap = TrimmableTypeMap.Instance;
 				var peer = typeMap.CreateInstance (reference.Handle, resolvedTargetType);
@@ -523,9 +105,16 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 				//      InvalidCastException via its `??` clause)
 				//  (c) classes are compatible but no proxy / activation failed
 				//      → NotSupportedException (genuine generator gap)
-				if (resolvedTargetType is not null &&
-						IsIncompatibleCast (typeMap, ref reference, resolvedTargetType)) {
-					return null;
+				if (targetType is not null && resolvedTargetType is not null) {
+					if (!typeMap.TryGetJniNameForManagedType (resolvedTargetType, out var targetJniName)) {
+						throw new ArgumentException (
+							$"Could not determine Java type corresponding to '{targetType.AssemblyQualifiedName}'.",
+							nameof (targetType));
+					}
+
+					if (JavaMarshalValueManagerHelper.IsIncompatibleCast (targetJniName, ref reference, resolvedTargetType)) {
+						return null;
+					}
 				}
 
 				var targetName = resolvedTargetType?.AssemblyQualifiedName ?? "<null>";
@@ -541,60 +130,6 @@ class JavaMarshalValueManager : AndroidReflectionJniValueManager
 		}
 
 		return base.CreatePeer (ref reference, transfer, targetType);
-	}
-
-	[return: DynamicallyAccessedMembers (Constructors)]
-	static Type? ResolvePeerType ([DynamicallyAccessedMembers (Constructors)] Type? type)
-	{
-		if (type is null) {
-			return null;
-		}
-		if (type == typeof (object) || type == typeof (IJavaPeerable)) {
-			return typeof (global::Java.Interop.JavaObject);
-		}
-		if (type == typeof (Exception)) {
-			return typeof (JavaException);
-		}
-		return type;
-	}
-
-	/// <summary>
-	/// Returns true when <paramref name="targetType"/>'s Java class is not assignable from
-	/// <paramref name="reference"/>. Throws when <paramref name="targetType"/> has no usable mapping.
-	/// </summary>
-	static bool IsIncompatibleCast (
-			TrimmableTypeMap typeMap,
-			ref JniObjectReference reference,
-			Type targetType)
-	{
-		if (!typeMap.TryGetJniNameForManagedType (targetType, out var targetJniName)) {
-			throw new ArgumentException (
-				$"Could not determine Java type corresponding to '{targetType.AssemblyQualifiedName}'.",
-				nameof (targetType));
-		}
-
-		var instanceClass = JniEnvironment.Types.GetObjectClass (reference);
-		JniObjectReference targetClass = default;
-		try {
-			try {
-				targetClass = JniEnvironment.Types.FindClass (targetJniName);
-			} catch (Java.Lang.ClassNotFoundException e) {
-				throw new ArgumentException (
-					$"Could not find Java class '{targetJniName}'.",
-					nameof (targetType), e);
-			}
-
-			if (!JniEnvironment.Types.IsAssignableFrom (instanceClass, targetClass)) {
-				// Bad cast: callers translate null to the expected result.
-				return true;
-			}
-		} finally {
-			JniObjectReference.Dispose (ref instanceClass);
-			JniObjectReference.Dispose (ref targetClass);
-		}
-
-		// Compatible classes mean a proxy/activation gap.
-		return false;
 	}
 
 	protected override bool TryConstructPeer (
